refactor: split pronunciation analysis
Split word, wordchunk and phoneme from pronunciation analysis into
their own files.
