+++ administrative-sdk/models/phoneme.js
@@ -1,0 +1,30 @@
+/**
+ * @class Phoneme
+ *
+ * @member {string} ipa The pronunciation of the grapheme(s) indicated as International Phonetic Alphabet (IPA).
+ * @member {float} score The audio is scored per phoneme and consists of several measurements. 0 would be bad,
+ * 1 the perfect score.
+ * @member {string} bad when the score is below 0.4, moderate when equal to 0.4 or between 0.4 and 0.6.
+ * good when the score is 0.6 or above.
+ */
+module.exports = class Phoneme {
+  /**
+   * Create a phoneme domain model.
+   *
+   * @constructor
+   * @param {string} ipa The pronunciation of the grapheme(s) indicated as International Phonetic Alphabet (IPA).
+   * @param {float} score The audio is scored per phoneme and consists of several measurements. 0 would be bad,
+   * 1 the perfect score.
+   * @param {float} confidenceScore This value provides a reliable prediction that the pronounced phoneme is
+   * actually the phoneme that is supposed to be pronounced. There is no absolute scale defined yet.
+   * @param {string} verdict bad when the score is below 0.4, moderate when equal to 0.4 or between 0.4 and 0.6.
+   * good when the score is 0.6 or above.
+   * @return {Phoneme}
+   */
+  constructor(ipa, score, confidenceScore, verdict) {
+    this.ipa = ipa;
+    this.score = score;
+    this.confidenceScore = confidenceScore;
+    this.verdict = verdict;
+  }
+};
